api/requirements.txt
gspread==6.0.0
google-auth==2.27.0
